requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
transformers
Pillow
torch
openai
gTTS
SpeechRecognition
pydub
streamlit-audiorec